+++ panels/gcode_macros.py
@@ -16,6 +16,7 @@ class MacroPanel(ScreenPanel):
     def initialize(self, panel_name):
         _ = self.lang.gettext
         self.macros = {}
+        self.loaded_macros = []
 
         # Create a scroll window for the macros
         scroll = Gtk.ScrolledWindow()
@@ -31,19 +32,24 @@ class MacroPanel(ScreenPanel):
         box.set_vexpand(True)
         box.pack_start(scroll, True, True, 0)
 
-        self.load_gcode_macros()
-
         self.content.add(box)
+
+    def activate(self):
+        self.unload_gcode_macros()
+        self.load_gcode_macros()
 
     def add_gcode_macro(self, macro):
 
         frame = Gtk.Frame()
         frame.set_property("shadow-type",Gtk.ShadowType.NONE)
+        frame.get_style_context().add_class("frame-item")
 
         name = Gtk.Label()
         name.set_markup("<big><b>%s</b></big>" % (macro))
         name.set_hexpand(True)
+        name.set_vexpand(True)
         name.set_halign(Gtk.Align.START)
+        name.set_valign(Gtk.Align.CENTER)
         name.set_line_wrap(True)
         name.set_line_wrap_mode(Pango.WrapMode.WORD_CHAR)
 
@@ -56,10 +62,6 @@ class MacroPanel(ScreenPanel):
         labels.add(name)
 
         dev = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-        dev.set_margin_top(10)
-        dev.set_margin_end(15)
-        dev.set_margin_start(15)
-        dev.set_margin_bottom(10)
         dev.set_hexpand(True)
         dev.set_vexpand(False)
 
@@ -74,6 +76,7 @@ class MacroPanel(ScreenPanel):
         macros = sorted(self.macros)
         pos = macros.index(macro)
 
+        self.loaded_macros.append(macro)
         self.labels['macros'].insert_row(pos)
         self.labels['macros'].attach(self.macros[macro]['row'], 0, pos, 1, 1)
         self.labels['macros'].show_all()
@@ -81,7 +84,28 @@ class MacroPanel(ScreenPanel):
     def load_gcode_macros(self):
         macros = self._screen.printer.get_gcode_macros()
         for x in macros:
-            self.add_gcode_macro(x[12:])
+            macro = x[12:].strip()
+
+            if macro in self.loaded_macros:
+                continue
+
+            logger.debug("Evaluating '%s' value '%s'" % (macro.strip().lower(),
+                self._config.get_config().getboolean("displayed_macros", macro.lower(), fallback=True)))
+
+            if ("displayed_macros" not in self._config.get_config().sections() or
+                    self._config.get_config().getboolean("displayed_macros", macro.lower(), fallback=True)):
+                self.add_gcode_macro(macro)
 
     def run_gcode_macro(self, widget, macro):
         self._screen._ws.klippy.gcode_script(macro)
+
+    def unload_gcode_macros(self):
+        for macro in self.loaded_macros:
+            logger.debug("Evaluating '%s' value '%s'" % (macro.strip().lower(),
+                self._config.get_config().getboolean("displayed_macros", macro.lower(), fallback=True)))
+            if ("displayed_macros" in self._config.get_config().sections() and
+                    not self._config.get_config().getboolean("displayed_macros", macro.lower(), fallback=True)):
+                macros = sorted(self.macros)
+                pos = macros.index(macro)
+                self.labels['macros'].remove_row(pos)
+                self.labels['macros'].show_all()