+++ panels/main_menu.py
@@ -28,16 +28,17 @@ class MainPanel(MenuPanel):
         for x in self._printer.get_tools():
             if i > 3:
                 break
-            self.labels[x] = self._gtk.ButtonImage("extruder-"+str(i+1), self._gtk.formatTemperatureString(0, 0))
-            eq_grid.attach(self.labels[x], i%2, i/2, 1, 1)
+            self.labels[x] = self._gtk.ButtonImage("extruder-"+str(i), self._gtk.formatTemperatureString(0, 0))
+            col = 0 if len(self._printer.get_tools()) == 1 else i%2 
+            row = i/2
+            eq_grid.attach(self.labels[x], col, row, 1, 1)
             i += 1
 
-        self.labels['heater_bed'] = self._gtk.ButtonImage("bed", self._gtk.formatTemperatureString(0, 0))
+        if self._printer.has_heated_bed():
+            self.labels['heater_bed'] = self._gtk.ButtonImage("bed", self._gtk.formatTemperatureString(0, 0))
 
-        width = 2 if i > 1 else 1
-        eq_grid.attach(self.labels['heater_bed'], 0, i/2+1, width, 1)
-
-        grid.attach(eq_grid, 0, 0, 1, 1)
+            width = 2 if i > 1 else 1
+            eq_grid.attach(self.labels['heater_bed'], 0, i/2+1, width, 1)
 
         self.items = items
         self.create_menu_items()
@@ -46,7 +47,9 @@ class MainPanel(MenuPanel):
         self.grid.set_row_homogeneous(True)
         self.grid.set_column_homogeneous(True)
 
+        grid.attach(eq_grid, 0, 0, 1, 1)
         grid.attach(self.arrangeMenuItems(items, 2, True), 1, 0, 1, 1)
+
         self.grid = grid
 
         self.target_temps = {