+++ screen.py
@@ -13,11 +13,13 @@ import importlib
 import logging
 import os
 import re
+import signal
 import subprocess
 
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk, Gdk, GLib, Pango
+from jinja2 import Environment, Template
 
 from ks_includes.KlippyWebsocket import KlippyWebsocket
 from ks_includes.KlippyRest import KlippyRest
@@ -54,12 +56,15 @@ class KlipperScreen(Gtk.Window):
     currentPanel = None
     files = None
     filename = ""
+    keyboard = None
+    keyboard_height = 200
     last_update = {}
     load_panel = {}
     number_tools = 1
     panels = {}
     popup_message = None
     printer = None
+    rtl_languages = ['he_il']
     subscriptions = []
     shutdown = True
 
@@ -75,8 +80,8 @@ class KlipperScreen(Gtk.Window):
         args = parser.parse_args()
         configfile = os.path.normpath(os.path.expanduser(args.configfile))
 
-
-        self._config = KlipperScreenConfig(configfile)
+        self.lang = gettext.translation('KlipperScreen', localedir='ks_includes/locales', fallback=True)
+        self._config = KlipperScreenConfig(configfile, self.lang, self)
         self.printer = Printer({
             "software_version": "Unknown"
         }, {
@@ -90,11 +95,35 @@ class KlipperScreen(Gtk.Window):
                 'is_active': False
             }
         })
-        self.lang = gettext.translation('KlipperScreen', localedir='ks_includes/locales', fallback=True)
+
+        self.printer.set_callbacks({
+            "disconnected": self.state_disconnected,
+            "error": self.state_error,
+            "printing": self.state_printing,
+            "ready": self.state_ready,
+            "startup": self.state_startup,
+            "shutdown": self.state_shutdown
+        })
+
+        logger.debug("OS Language: %s" % os.getenv('LANG'))
+
+        self.lang_ltr = True
+        for lang in self.rtl_languages:
+            if os.getenv('LANG').lower().startswith(lang):
+                self.lang_ltr = False
+                Gtk.Widget.set_default_direction(Gtk.TextDirection.RTL)
+                logger.debug("Enabling RTL mode")
+                break
+
         _ = self.lang.gettext
 
         self.apiclient = KlippyRest(self._config.get_main_config_option("moonraker_host"),
-            self._config.get_main_config_option("moonraker_port"))
+            self._config.get_main_config_option("moonraker_port"),
+            self._config.get_main_config_option("moonraker_api_key", False))
+
+        powerdevs = self.apiclient.send_request("machine/device_power/devices")
+        if powerdevs != False:
+            self.printer.configure_power_devices(powerdevs['result'])
 
         Gtk.Window.__init__(self)
         self.width = self._config.get_main_config().getint("width", Gdk.Screen.get_width(Gdk.Screen.get_default()))
@@ -119,25 +148,26 @@ class KlipperScreen(Gtk.Window):
             self._config.get_main_config_option("moonraker_host"),
             self._config.get_main_config_option("moonraker_port")
         )
-        self._ws.connect()
+        self._ws.initial_connect()
 
         # Disable DPMS
-        os.system("/usr/bin/xset -display :0 s off")
         os.system("/usr/bin/xset -display :0 -dpms")
-        os.system("/usr/bin/xset -display :0 s noblank")
+        self.set_screenblanking_timeout(self._config.get_main_config_option('screen_blanking'))
+
+        # Change cursor to blank
+        self.get_window().set_cursor(Gdk.Cursor(Gdk.CursorType.BLANK_CURSOR))
 
         return
-
 
     def ws_subscribe(self):
         requested_updates = {
             "objects": {
                 "bed_mesh": ["profile_name","mesh_max","mesh_min","probed_matrix"],
                 "configfile": ["config"],
-                "extruder": ["target","temperature","pressure_advance","smooth_time"],
                 "fan": ["speed"],
                 "gcode_move": ["extrude_factor","gcode_position","homing_origin","speed_factor"],
                 "heater_bed": ["target","temperature"],
+                "idle_timeout": ["state"],
                 "pause_resume": ["is_paused"],
                 "print_stats": ["print_duration","total_duration","filament_used","filename","state","message"],
                 "toolhead": ["homed_axes","estimated_print_time","print_time","position","extruder"],
@@ -145,6 +175,9 @@ class KlipperScreen(Gtk.Window):
                 "webhooks": ["state","state_message"]
             }
         }
+        for extruder in self.printer.get_tools():
+            requested_updates['objects'][extruder] = ["target","temperature","pressure_advance","smooth_time"]
+
         self._ws.klippy.object_subscription(requested_updates)
 
     def _load_panel(self, panel, *args):
@@ -171,7 +204,6 @@ class KlipperScreen(Gtk.Window):
             logger.exception(msg)
             raise Exception(msg)
 
-
     def show_panel(self, panel_name, type, title, remove=None, pop=True, **kwargs):
         if panel_name not in self.panels:
             self.panels[panel_name] = self._load_panel(type, self, title)
@@ -183,14 +215,12 @@ class KlipperScreen(Gtk.Window):
                     self.panels[panel_name].initialize(panel_name)
             except:
                 del self.panels[panel_name]
+                logger.exception("Unable to load panel %s" % type)
                 self.show_error_modal("Unable to load panel %s" % type)
                 return
 
             if hasattr(self.panels[panel_name],"process_update"):
                 self.panels[panel_name].process_update("notify_status_update", self.printer.get_data())
-
-        if hasattr(self.panels[panel_name],"activate"):
-            self.panels[panel_name].activate()
 
         if remove == 2:
             self._remove_all_panels()
@@ -199,6 +229,11 @@ class KlipperScreen(Gtk.Window):
 
         self.add(self.panels[panel_name].get())
         self.show_all()
+
+        if hasattr(self.panels[panel_name],"activate"):
+            self.panels[panel_name].activate()
+            self.show_all()
+
         self._cur_panels.append(panel_name)
         logger.debug("Current panel hierarchy: %s", str(self._cur_panels))
 
@@ -297,8 +332,10 @@ class KlipperScreen(Gtk.Window):
 
         # Find current menu item
         panels = list(self._cur_panels)
-        if "job_status" not in self._cur_panels:
+        if "main_panel" in self._cur_panels:
             menu = "__main"
+        elif "splash_screen" in self._cur_panels:
+            menu = "__splashscreen"
         else:
             menu = "__print"
 
@@ -329,10 +366,12 @@ class KlipperScreen(Gtk.Window):
 
     def _menu_go_back (self, widget=None):
         logger.info("#### Menu go back")
+        self.remove_keyboard()
         self._remove_current_panel()
 
     def _menu_go_home(self):
         logger.info("#### Menu go home")
+        self.remove_keyboard()
         while len(self._cur_panels) > 1:
             self._remove_current_panel()
 
@@ -350,34 +389,71 @@ class KlipperScreen(Gtk.Window):
                 self.subscriptions.pop(i)
                 return
 
+    def set_screenblanking_timeout(self, time):
+        logger.debug("Changing screenblanking to: %s" % time)
+        if time == "off":
+            os.system("/usr/bin/xset -display :0 s off")
+            os.system("/usr/bin/xset -display :0 s noblank")
+            return
+
+        time = int(time)
+        if time < 0:
+            return
+
+        os.system("/usr/bin/xset -display :0 s on")
+        os.system("/usr/bin/xset -display :0 s %s" % time)
+
+    def state_disconnected(self):
+        _ = self.lang.gettext
+        logger.debug("### Going to disconnected")
+        self.printer_initializing(_("Klipper has disconnected"))
+
+    def state_error(self):
+        _ = self.lang.gettext
+        msg = self.printer.get_stat("webhooks","state_message")
+        if "FIRMWARE_RESTART" in msg:
+            self.printer_initializing(
+                _("Klipper has encountered an error.\nIssue a FIRMWARE_RESTART to attempt fixing the issue.")
+            )
+        elif "micro-controller" in msg:
+            self.printer_initializing(
+                _("Klipper has encountered an error with the micro-controller.\nPlease recompile and flash.")
+            )
+        else:
+            self.printer_initializing(
+                _("Klipper has encountered an error.")
+            )
+
+    def state_printing(self):
+        if "job_status" in self._cur_panels:
+            return
+        self.printer_printing()
+
+    def state_ready(self):
+        # Do not return to main menu if completing a job, timeouts/user input will return
+        if "job_status" in self._cur_panels or "main_menu" in self._cur_panels:
+            return
+        self.printer_ready()
+
+    def state_startup(self):
+        _ = self.lang.gettext
+        self.printer_initializing(_("Klipper is attempting to start"))
+
+    def state_shutdown(self):
+        _ = self.lang.gettext
+        self.printer_initializing(_("Klipper has shutdown"))
+
     def _websocket_callback(self, action, data):
         _ = self.lang.gettext
-        #print(json.dumps([action, data], indent=2))
 
         if action == "notify_klippy_disconnected":
-            logger.info("### Going to disconnected state")
-            self.printer_initializing(_("Klipper has shutdown"))
+            logger.debug("Received notify_klippy_disconnected")
+            self.printer.change_state("disconnected")
             return
         elif action == "notify_klippy_ready":
-            logger.info("### Going to ready state")
-            self.init_printer()
-        elif action == "notify_status_update" and self.shutdown == False:
+            self.printer.change_state("ready")
+        elif action == "notify_status_update" and self.printer.get_state() != "shutdown":
             self.printer.process_update(data)
-            if "webhooks" in data:
-                print(json.dumps([action, data], indent=2))
-            if "webhooks" in data and "state" in data['webhooks']:
-                if data['webhooks']['state'] == "ready":
-                    logger.info("### Going to ready state")
-                    self.printer_ready()
-                elif data['webhooks']['state'] == "shutdown":
-                    self.shutdown == True
-                    self.printer_initializing(_("Klipper has shutdown"))
-            else:
-                active = self.printer.get_stat('virtual_sdcard','is_active')
-                paused = self.printer.get_stat('pause_resume','is_paused')
-                if "job_status" not in self._cur_panels:
-                    if active == True or paused == True:
-                        self.printer_printing()
         elif action == "notify_filelist_changed":
             logger.debug("Filelist changed: %s", json.dumps(data,indent=2))
             #self.files.add_file()
@@ -386,10 +462,10 @@ class KlipperScreen(Gtk.Window):
         elif action == "notify_power_changed":
             logger.debug("Power status changed: %s", data)
             self.printer.process_power_update(data)
-        elif self.shutdown == False and action == "notify_gcode_response":
+        elif self.printer.get_state() not in ["error","shutdown"] and action == "notify_gcode_response":
             if "Klipper state: Shutdown" in data:
-                self.shutdown == True
-                self.printer_initializing(_("Klipper has shutdown"))
+                logger.debug("Shutdown in gcode response, changing state to shutdown")
+                self.printer.change_state("shutdown")
 
             if not (data.startswith("B:") and
                 re.search(r'B:[0-9\.]+\s/[0-9\.]+\sT[0-9]+:[0-9\.]+', data)):
@@ -400,13 +476,21 @@ class KlipperScreen(Gtk.Window):
         for sub in self.subscriptions:
             self.panels[sub].process_update(action, data)
 
-    def _confirm_send_action(self, widget, text, method, params):
+    def _confirm_send_action(self, widget, text, method, params={}):
         _ = self.lang.gettext
 
         buttons = [
             {"name":_("Continue"), "response": Gtk.ResponseType.OK},
             {"name":_("Cancel"),"response": Gtk.ResponseType.CANCEL}
         ]
+
+        try:
+            env = Environment(extensions=["jinja2.ext.i18n"])
+            env.install_gettext_translations(self.lang)
+            j2_temp = env.from_string(text)
+            text = j2_temp.render()
+        except:
+            logger.debug("Error parsing jinja for confirm_send_action")
 
         label = Gtk.Label()
         label.set_markup(text)
@@ -436,7 +520,6 @@ class KlipperScreen(Gtk.Window):
 
     def init_printer(self):
         _ = self.lang.gettext
-        self.shutdown = False
 
         status_objects = [
             'bed_mesh',
@@ -449,73 +532,87 @@ class KlipperScreen(Gtk.Window):
             'print_stats',
             'heater_bed',
             'extruder',
-            'pause_resume'
+            'pause_resume',
+            'webhooks'
         ]
         printer_info = self.apiclient.get_printer_info()
+        logger.debug("Sending request %s" % "printer/objects/query?" + "&".join(status_objects))
         data = self.apiclient.send_request("printer/objects/query?" + "&".join(status_objects))
         powerdevs = self.apiclient.send_request("machine/device_power/devices")
-        if printer_info == False or data == False:
-            self.printer_initializing(_("Moonraker error"))
-            return
         data = data['result']['status']
-
-        # Reinitialize printer, in case the printer was shut down and anything has changed.
-        self.printer.__init__(printer_info['result'], data)
-        self.ws_subscribe()
-
-        if powerdevs != False:
-            self.printer.configure_power_devices(powerdevs['result'])
 
         if self.files == None:
             self.files = KlippyFiles(self)
         else:
             self.files.add_timeout()
 
-        if printer_info['result']['state'] in ("error","shutdown","startup"):
-            if printer_info['result']['state'] == "startup":
-                self.printer_initializing(_("Klipper is attempting to start"))
-            elif printer_info['result']['state'] == "error":
-                if "FIRMWARE_RESTART" in printer_info['result']['state_message']:
-                    self.printer_initializing(
-                        _("Klipper has encountered an error.\nIssue a FIRMWARE_RESTART to attempt fixing the issue.")
-                    )
-                elif "micro-controller" in printer_info['result']['state_message']:
-                    self.printer_initializing(
-                        _("Klipper has encountered an error with the micro-controller.\nPlease recompile and flash.")
-                    )
-                else:
-                    self.printer_initializing(
-                        _("Klipper has encountered an error.")
-                    )
-            else:
-                self.printer_initializing(_("Klipper has shutdown"))
-            return
-        if (data['print_stats']['state'] == "printing" or data['print_stats']['state'] == "paused"):
-            filename = self.printer.get_stat("print_stats","filename")
-            if not self.files.file_metadata_exists(filename):
-                self.files.request_metadata(filename)
-            self.printer_printing()
-            return
-        self.printer_ready()
+        # Reinitialize printer, in case the printer was shut down and anything has changed.
+        self.printer.reinit(printer_info['result'], data)
+        self.ws_subscribe()
+
+        if powerdevs != False:
+            self.printer.configure_power_devices(powerdevs['result'])
 
     def printer_ready(self):
-        if self.shutdown == True:
-            self.init_printer()
-            return
-
-        self.files.add_timeout()
         self.close_popup_message()
+        # Force update to printer webhooks state in case the update is missed due to websocket subscribe not yet sent
+        self.printer.process_update({"webhooks":{"state":"ready","state_message": "Printer is ready"}})
         self.show_panel('main_panel', "main_menu", "Main Menu", 2, items=self._config.get_menu_items("__main"),
             extrudercount=self.printer.get_extruder_count())
+        self.ws_subscribe()
         if "job_status" in self.panels:
             self.remove_subscription("job_status")
             del self.panels["job_status"]
 
     def printer_printing(self):
-        self.ws_subscribe()
         self.files.remove_timeout()
         self.close_popup_message()
         self.show_panel('job_status',"job_status", "Print Status", 2)
+
+    def show_keyboard(self, widget=None):
+        if self.keyboard is not None:
+            return
+
+        env = os.environ.copy()
+        env["MB_KBD_CONFIG"] = "/home/pi/.matchbox/keyboard.xml"
+        env["MB_KBD_CONFIG"] = "ks_includes/locales/keyboard.xml"
+        p = subprocess.Popen(["matchbox-keyboard", "--xid"], stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE, env=env)
+        #p = subprocess.Popen(["onboard", "--xid"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        xid = int(p.stdout.readline())
+        logger.debug("XID %s" % xid)
+        logger.debug("PID %s" % p.pid)
+        keyboard = Gtk.Socket()
+        #keyboard.connect("plug-added", self.plug_added)
+        box = Gtk.VBox()
+        box.set_size_request(self.width, self.keyboard_height)
+        box.add(keyboard)
+
+        cur_panel = self.panels[self._cur_panels[-1]]
+        #for i in ['back','estop','home']:
+        #    if i in cur_panel.control:
+        #        cur_panel.control[i].set_sensitive(False)
+        cur_panel.get().put(box, 0, self.height - 200)
+        self.show_all()
+        keyboard.add_id(xid)
+        keyboard.show()
+
+        self.keyboard = {
+            "box": box,
+            "panel": cur_panel.get(),
+            "process": p,
+            "socket": keyboard
+        }
+
+
+    def remove_keyboard(self, widget=None):
+        if self.keyboard is None:
+            return
+
+        self.keyboard['panel'].remove(self.keyboard['box'])
+        os.kill(self.keyboard['process'].pid, signal.SIGTERM)
+        self.keyboard = None
+
 
 def get_software_version():
     prog = ('git', '-C', os.path.dirname(__file__), 'describe', '--always',